Update versions.cfg
I had a probleme with the RelatedField which does not work if eea.facetednavigation is installed. the Even after uninstalling eea.facetednavigation it was not working because collective.js.jqueryui.custom.min.js was not removed from the js registry. With the current version of eea.facetednavigation the RelationField works.